Added String-repeatify-method.js which extends String object
The exercise demonstrate inheritance in JavaScript and the prototype property. It also verifies that the developer is able to extend native data type functionalities (although this should not be done). Another important point here is to demonstrate that you are aware about how to not override possible already defined functions.